fix: limit에 맞는 데이터 타입으로 교체

+++ src/main/java/com/maps/coin/dto/CreateRoomRequest.java
@@ -10,8 +10,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class CreateRoomRequest {
     private String name;
-    private Long personnel;
-    private Long size;
+    private Integer personnel;
+    private Integer size;
 
     public Room toEntity() {
         return Room.builder()
